Added Int64

+++ Sources/Apio/Generate.swift
@@ -906,8 +906,8 @@ func generateResultValueType(valueType: ResultValueType) -> String
             return "String"
         case .identifier:
             return "String"
-        case .int:
-            return "Int"
+        case .int64:
+            return "Int64"
     }
 }
